+++ R/abnormal_by_worst_grade.R
@@ -101,7 +101,8 @@ s_count_abnormal_by_worst_grade <- function(df, # nolint
   assert_valid_factor(df[[variables$param]])
   assert_valid_factor(df[[variables$grade_dir]])
   assert_df_with_variables(df, c(a = .var, variables))
-  assert_character_or_factor(df[[variables$id]])
+  checkmate::assertMultiClass(df[[variables$id]], classes = c("factor", "character"))
+
   # To verify that the `split_rows_by` are performed with correct variables.
   assertthat::assert_that(
     all(