+++ R/summarize_num_patients.R
@@ -39,7 +39,7 @@ s_num_patients <- function(x, labelstr, .N_col, count_by = NULL) { # nolint
     assertthat::is.string(labelstr)
   )
   checkmate::assert_count(.N_col)
-  assert_character_or_factor(x)
+  checkmate::assertMultiClass(x, classes = c("factor", "character"))
 
   count1 <- n_available(unique(x))
   count2 <- n_available(x)
@@ -48,7 +48,7 @@ s_num_patients <- function(x, labelstr, .N_col, count_by = NULL) { # nolint
     assertthat::assert_that(
       is_equal_length(count_by, x)
     )
-    assert_character_or_factor(count_by)
+    checkmate::assertMultiClass(count_by, classes = c("factor", "character"))
     count2 <- n_available(unique(interaction(x, count_by)))
   }