+++ tests/testthat/test-utils_checkmate.R
@@ -1,16 +1,3 @@
-# assert_character_or_factor ----
-
-testthat::test_that("assert_character_or_factor is TRUE with healthy input", {
-  testthat::expect_silent(assert_character_or_factor(c("a", "b")))
-  testthat::expect_silent(assert_character_or_factor(factor(c("a", "b"))))
-  testthat::expect_silent(assert_character_or_factor(NA_character_))
-})
-
-testthat::test_that("assert_character_or_factor is FALSE with wrong input", {
-  testthat::expect_error(assert_character_or_factor(c(5L, 3L)))
-  testthat::expect_error(assert_character_or_factor(NULL))
-})
-
 # assert_list_of_variables ----
 
 testthat::test_that("assert_list_of_variables is TRUE with healthy input", {