+++ NEWS.md
@@ -1,8 +1,7 @@
 # tern 0.7.8.9006
 
 ### Migration from `assertthat` to `checkmate`
-* `is_character_or_factor` is now `assert_character_or_factor` and uses 
-  `checkmate::assertMultiClass()` and made internal.
+* `is_character_or_factor` replaced by `checkmate::assertMultiClass()`.
 * `is_nonnegative_count` removed and substituted by `checkmate::assert_count()`.
 * `assert_list_of_variables` instead of `is_variables` (made internal).
 * `assert_df_with_variables` instead of `is_df_with_variables` (made internal).